Passed Wave 1

+++ app/video_routes.py
@@ -1,11 +1,9 @@
 
 from flask import Blueprint
-from app.models.customer import Customer
 from app.models.video import Video
 from app import db
-from flask import request, Blueprint, make_response, Response, jsonify
-from datetime import date
-import requests
+from flask import request, Blueprint, make_response, jsonify
+
 
 video_bp = Blueprint("videos", __name__, url_prefix="/videos")
 @video_bp.route("", methods=["GET", "POST"])
@@ -17,31 +15,17 @@ def list_videos():  # NameError
         videos_response = []
 
         for video in videos:
-            videos_response.append({
-                "id": video.video_id,
-                "title": video.title,
-                "release_date": video.release_date,
-                "total_inventory":video.total_inventory,
-                "availiable_inventory":video.availiable_inventory
-                }, 
-                {
-                "id": video.video_id,
-                "title": video.title,
-                "release_date": video.release_date,
-                "total_inventory":video.total_inventory,
-                "availiable_inventory":video.availiable_inventory 
-                })
-            #The API should return an empty array and a status 200 if there are no videos.
-        return jsonify(videos_response)
+            videos_response.append(video.video_get_json())
+        return jsonify(videos_response), 200
 
 # POST /videos
     elif request.method == "POST":  # CRUD CREATE
         # check for request body title and description, plus ensure both are strings
         request_body = request.get_json()
 
-        if "title" or "release_date" or "total_inventory" not in request_body:
+        if "title" not in request_body or "release_date" not in request_body or "total_inventory" not in request_body:
             return {
-                "details": "Bad Request"
+                "details": "Bad Request"    #correct?
             }, 400
 
         video = Video(
@@ -53,11 +37,7 @@ def list_videos():  # NameError
         db.session.add(video)
         db.session.commit()
 
-        return {
-            "video": {
-                "id": video.video_id
-            }
-        }, 201
+        return video.video_get_json(), 201 #where video is created?
 
 @video_bp.route("/<video_id>", methods=["GET", "PUT", "DELETE"])
 def single_video(video_id):
@@ -68,39 +48,40 @@ def single_video(video_id):
 
 # GET /vidoes/<id>
     if request.method == "GET":
-        return {
-            "video": {
-                "id": video.video_id,
-                "title": video.title,
-                "release_date": video.release_date,
-                "total_inventory":video.total_inventory,
-                "availiable_inventory":video.availiable_inventory
-            }
-        },200
+        return video.video_get_json(),200
 
 # PUT /videos/<id>
     elif request.method == "PUT":
-        form_data = request.get_json()
+        request_body = request.get_json()
 
-        video.title = form_data["title"]
+        video = Video.query.get(video_id)
+        if video == None:
+            return make_response(f"Video {video_id} not found"), 404
 
+        if "title" not in request_body or "release_date" not in request_body or "total_inventory" not in request_body:
+            return {
+                "details": f"Not found"
+            }, 400
+
+        #update customer info
+        video.title = request_body["title"]
+        video.release_date= request_body["release_date"]
+        video.total_inventory= request_body["total_inventory"]
+        
         db.session.commit()
 
-        return {
-            "video": {
-                "id": video.video_id,
-                "title": video.title,
-                "release_date": video.release_date,
-                "total_inventory":video.total_inventory,
-                "availiable_inventory":video.availiable_inventory
-                }
-            }, 200
+        return video.video_get_json(), 200
     
 # DELETE /videos/<id>
     elif request.method == "DELETE":
+        video = Video.query.get(video_id)
+        if video == None:
+            return make_response(f"Video {video_id} not found", 404)
+        
         db.session.delete(video)
         db.session.commit()
-            
+
         return {
-            "details": f"Video {video.video_id} \"{video.title}\" successfully deleted"
-        },200
+            "id":int(video_id)
+        }, 200
+
